Get the privacy toggle working again
This thing is a bit of a beast.

In order to make a note private at the moment we move it into a different directory inside either
the public or private directories in the user's POD. We do this because we don't want to muck around with
ACLs quite yet, but this design decision should be revisited soon.

For the moment, this means that we need to do several operations when we change privacy:

1) create the new note resource
2) update both the public and private index and finally
3) delete the old note resource

This is why we have so many hooks below, and two separate functions for making a note public or private.