Use normalizer for mentions too
I was going to try to abstract this, but it
due to the way normalizers work (specifically the need to return after a single action) the abstraction doesn't buy us much.

This may get simpler once we feel comortable removing the migration logic from the normalizer, and I suspect we'll find a way to abstract some of this at some point, but for now I think the code duplication is justified.