get toolbars working
I think the upgrade to Plate 10 actually broke the toolbars in a subtle way - weirdly it looks like if we import BlockToolbarButton or MarkToolbarButton from plate they can't find the plate context we find - I'm assuming this has something to do with module imports (maybe we're using a module that stores provider state in a different place? global state is fun!) so tried to clean up a few module imports, but this is still a bit mysterious.

to solve, I copied {Block| Mark}ToolbarButtons from plate and tweaked them a bit - not so bad to have this stuff in our codebase anyway as we may want to tweak them in our own ways soon